create controller with resources

+++ app/Http/Controllers/Country/CountryController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
 use App\Models\CountryModel;
+// use Validator;
 
 class CountryController extends Controller
 {
@@ -14,22 +15,50 @@ class CountryController extends Controller
     }
 
     public function countryById($id){
-        return response()->json(CountryModel::find($id), 200);
+        $country = CountryModel::find($id);
+
+        if (is_null($country)) {
+            return response()->json("Record not found.!", 404);
+        }
+        return response()->json($country, 200);
     }
 
     public function countrySave(Request $request){
+        // $validator = $request->validate([
+        //     'name' => 'required|min:3',
+        //     'iso' => 'required|min:2|max:2'
+        // ]);
+
+        // if ($validator->)
+
+
+
         $country = CountryModel::create($request->all());
 
         return response()->json($country, 201);
     }
 
-    public function countryUpdate(Request $request, CountryModel $country){
+    public function countryUpdate(Request $request, $id)
+    {
+        $country = CountryModel::find($id);
+
+        if (is_null($country)) {
+            return response()->json("Record not found.!", 404);
+        }
+
         $country->update($request->all());
 
         return response()->json($country, 200);
     }
 
-    public function countryDelete(Request $request, CountryModel $country){
+    public function countryDelete($id)
+    {
+        $country = CountryModel::find($id);
+
+        if (is_null($country)) {
+            return response()->json("Record not found.!", 404);
+        }
+
         $country->delete();
 
         return response()->json(null, 204);
